:art: move GetFilmDetailsService's send method into parent Service class
